Create game structure
The game was developed using the "p5.js" platform, which is an animation library created based on the JavaScript programming language. Pong is a two-dimensional sports electronic game that simulates a table tennis match. The player takes control of a vertical paddle located on the left side of the screen and competes against the computer or another player, who controls a second paddle on the opposite side.